Move and cleanup
