+++ include/aws/io/tls_channel_handler.h
@@ -486,6 +486,17 @@ AWS_IO_API struct aws_byte_buf aws_tls_handler_protocol(struct aws_channel_handl
  */
 AWS_IO_API struct aws_byte_buf aws_tls_handler_server_name(struct aws_channel_handler *handler);
 
+/********************************* Misc TLS related *********************************/
+
+/*
+ * Injects a tls handler/slot into a channel and begins tls negotiation.
+ * If desired, ALPN must be handled separately
+ */
+AWS_IO_API int aws_channel_setup_client_tls(
+    struct aws_channel *channel,
+    struct aws_allocator *allocator,
+    struct aws_tls_connection_options *tls_options);
+
 AWS_EXTERN_C_END
 
 #endif /* AWS_IO_TLS_CHANNEL_HANDLER_H */
